refactor: move send_batch_inclusion_data_responses to connection
